requirements.txt
torch~=1.13.0
numpy~=1.23.0
pandas~=2.1.4
PyYAML~=6.0.1
tqdm~=4.65.0
h5py~=3.10.0
geojson~=3.1.0
networkx~=3.2.1
matplotlib~=3.8.2
geopandas~=0.14.3
shapely~=2.0.3
geopy~=2.4.1
scikit-image~=0.22.0